fix(exec): share control command approval guards
